Added Vital Signs and Diagnosis API

+++ src/main/java/com/pms/dto/IdDto.java
@@ -2,13 +2,13 @@ package com.pms.dto;
 
 public class IdDto {
 	
-	public Integer id;
+	public Long id;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
